Urgenti
-aggiustare le add e gli update

-per ogni update controllare che la lista non sia vuota e fare i link

-risolvere bug che non collega pagamento e tipo pagamento nel db

Importante

-modificare le responseEntity in caso di errore in modo che ritornino il messaggio di errore

Risolti
-creare ordine parametri: lista prodotti, idUser
-risolvi categoria_id null, in particolare gestisci link sia prodotto -> categoria sia categoria -> prodotti (crea un prodotto ex novo e aggiungi categoria prima di aggiungere al db), nel DTO quando fai il .of dall'entity controlla che esista l'oggetto della chiave esterna
-le date le inserisci tu, non sono necessarie da richiedere
-endpoint, return feedback, proposte, ordini, prodotti per ogni utente
-gestisci che ad ogni ordine rimuovi dal prodotto la quantità ordinata, te la becchi dalle informazioni di ogni rigaOrdine
-import statico dell'enum dell'exception

+++ EshopELIS/src/main/java/com/elismarket/eshop/eshopelis/controller/ProdottoController.java
@@ -47,7 +47,7 @@ public class ProdottoController {
     }
 
     @GetMapping("/all/utente/{userId}")
-    public List<ProdottoDTO> findAllByUtente(@PathVariable Integer userId) {
+    public List<ProdottoDTO> findAllByUtente(@PathVariable Long userId) {
         return prodottoService.findAllByUtente(userId);
     }
